Separate REPL-related cases from main reducer
