+++ src/commons/workspace/WorkspaceReducer.ts
@@ -17,7 +17,6 @@ import {
   handleConsoleLog
 } from '../application/actions/InterpreterActions';
 import {
-  CodeOutput,
   createDefaultWorkspace,
   defaultWorkspaceManager,
   ErrorOutput,
@@ -32,33 +31,25 @@ import {
 } from '../collabEditing/CollabEditingActions';
 import { NOTIFY_PROGRAM_EVALUATED } from '../sideContent/SideContentTypes';
 import { SourceActionType } from '../utils/ActionsHelper';
-import Constants from '../utils/Constants';
 import { createContext } from '../utils/JsSlangHelper';
 import { handleCseAndStepperActions } from './reducers/cseReducer';
 import { handleEditorActions } from './reducers/editorReducer';
+import { handleReplActions } from './reducers/replReducer';
 import {
-  browseReplHistoryDown,
-  browseReplHistoryUp,
   changeExecTime,
   changeExternalLibrary,
-  clearReplInput,
-  clearReplOutput,
-  clearReplOutputLast,
   disableTokenCounter,
   enableTokenCounter,
   endClearContext,
   evalEditor,
-  evalRepl,
   resetTestcase,
   resetWorkspace,
-  sendReplInputToOutput,
   setIsEditorReadonly,
   setTokenCount,
   toggleEditorAutorun,
   updateCurrentAssessmentId,
   updateCurrentSubmissionId,
   updateHasUnsavedChanges,
-  updateReplValue,
   updateSublanguage,
   updateSubmissionsTableFilters,
   updateWorkspace
@@ -119,86 +110,15 @@ export const WorkspaceReducer: Reducer<WorkspaceManagerState, SourceActionType> 
 const newWorkspaceReducer = createReducer(defaultWorkspaceManager, builder => {
   handleEditorActions(builder);
   handleCseAndStepperActions(builder);
+  handleReplActions(builder);
   builder
     .addCase(setTokenCount, (state, action) => {
       const workspaceLocation = getWorkspaceLocation(action);
       state[workspaceLocation].tokenCount = action.payload.tokenCount;
     })
-    .addCase(browseReplHistoryDown, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      if (state[workspaceLocation].replHistory.browseIndex === null) {
-        // Not yet started browsing history, nothing to do
-        return;
-      }
-      if (state[workspaceLocation].replHistory.browseIndex !== 0) {
-        // Browsing history, and still have earlier records to show
-        const newIndex = state[workspaceLocation].replHistory.browseIndex! - 1;
-        const newReplValue = state[workspaceLocation].replHistory.records[newIndex];
-
-        state[workspaceLocation].replValue = newReplValue;
-        state[workspaceLocation].replHistory.browseIndex = newIndex;
-        return;
-      }
-      // Browsing history, no earlier records to show; return replValue to
-      // the last value when user started browsing
-      const newIndex = null;
-      const newReplValue = state[workspaceLocation].replHistory.originalValue;
-      const newRecords = state[workspaceLocation].replHistory.records.slice();
-
-      state[workspaceLocation].replValue = newReplValue;
-      state[workspaceLocation].replHistory = {
-        browseIndex: newIndex,
-        records: newRecords,
-        originalValue: ''
-      };
-    })
-    .addCase(browseReplHistoryUp, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      const lastRecords = state[workspaceLocation].replHistory.records;
-      const lastIndex = state[workspaceLocation].replHistory.browseIndex;
-      if (
-        lastRecords.length === 0 ||
-        (lastIndex !== null && lastRecords[lastIndex + 1] === undefined)
-      ) {
-        // There is no more later history to show
-        return;
-      }
-      if (lastIndex === null) {
-        // Not yet started browsing, initialise the index & array
-        const newIndex = 0;
-        const newRecords = lastRecords.slice();
-        const originalValue = state[workspaceLocation].replValue;
-        const newReplValue = newRecords[newIndex];
-
-        state[workspaceLocation].replValue = newReplValue;
-        state[workspaceLocation].replHistory = {
-          browseIndex: newIndex,
-          records: newRecords,
-          originalValue
-        };
-        return;
-      }
-      // Browsing history, and still have later history to show
-      const newIndex = lastIndex + 1;
-      const newReplValue = lastRecords[newIndex];
-      state[workspaceLocation].replValue = newReplValue;
-      state[workspaceLocation].replHistory.browseIndex = newIndex;
-    })
     .addCase(changeExecTime, (state, action) => {
       const workspaceLocation = getWorkspaceLocation(action);
       state[workspaceLocation].execTime = action.payload.execTime;
-    })
-    .addCase(clearReplInput, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      state[workspaceLocation].replValue = '';
-    })
-    .addCase(clearReplOutputLast, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      state[workspaceLocation].output.pop();
-    })
-    .addCase(clearReplOutput, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      state[workspaceLocation].output = [];
     })
     .addCase(endClearContext, (state, action) => {
       const workspaceLocation = getWorkspaceLocation(action);
@@ -218,28 +138,6 @@ const newWorkspaceReducer = createReducer(defaultWorkspaceManager, builder => {
           externalLibrary: action.payload.library.external.name
         }
       };
-    })
-    .addCase(sendReplInputToOutput, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      // CodeOutput properties exist in parallel with workspaceLocation
-      const newOutput: InterpreterOutput[] = state[workspaceLocation].output.concat(
-        action.payload as CodeOutput
-      );
-
-      let newReplHistoryRecords: string[];
-      if (action.payload.value !== '') {
-        newReplHistoryRecords = [action.payload.value].concat(
-          state[workspaceLocation].replHistory.records
-        );
-      } else {
-        newReplHistoryRecords = state[workspaceLocation].replHistory.records;
-      }
-      if (newReplHistoryRecords.length > Constants.maxBrowseIndex) {
-        newReplHistoryRecords.pop();
-      }
-
-      state[workspaceLocation].output = newOutput;
-      state[workspaceLocation].replHistory.records = newReplHistoryRecords;
     })
     .addCase(changeExternalLibrary, (state, action) => {
       const workspaceLocation = getWorkspaceLocation(action);
@@ -292,10 +190,6 @@ const newWorkspaceReducer = createReducer(defaultWorkspaceManager, builder => {
       const workspaceLocation = getWorkspaceLocation(action);
       state[workspaceLocation].isRunning = true;
       state[workspaceLocation].isDebugging = false;
-    })
-    .addCase(evalRepl, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      state[workspaceLocation].isRunning = true;
     })
     .addCase(evalInterpreterSuccess, (state, action) => {
       const workspaceLocation = getWorkspaceLocation(action);
@@ -480,10 +374,6 @@ const newWorkspaceReducer = createReducer(defaultWorkspaceManager, builder => {
     .addCase(updateCurrentSubmissionId, (state, action) => {
       state.grading.currentSubmission = action.payload.submissionId;
       state.grading.currentQuestion = action.payload.questionId;
-    })
-    .addCase(updateReplValue, (state, action) => {
-      const workspaceLocation = getWorkspaceLocation(action);
-      state[workspaceLocation].replValue = action.payload.newReplValue;
     })
     .addCase(updateHasUnsavedChanges, (state, action) => {
       // For some reason mutating the state directly results in type
